set ttl explicitly in nsupdate rather than leaving it to inheritence.
somewhat longer explanation is that ansible default behavior changed;
there is a default 3600 minute ttl in the module which *IS ENFORCED*
rather than leaving it to inheriting from the zone.

This WILL result in a lot of head scratching and frustration when
debugging new hosts.  Our explicit TTL for the handshake records is now
a more reasonable 60 seconds.